+++ stride-application/src/main/java/io/descoped/stride/application/StrideApplicationImpl.java
@@ -5,24 +5,14 @@ import no.cantara.config.ApplicationProperties;
 import org.eclipse.jetty.server.Connector;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
-import org.glassfish.hk2.api.DynamicConfiguration;
-import org.glassfish.hk2.api.DynamicConfigurationService;
 import org.glassfish.hk2.api.Filter;
-import org.glassfish.hk2.api.MultiException;
-import org.glassfish.hk2.api.Populator;
 import org.glassfish.hk2.api.ServiceHandle;
 import org.glassfish.hk2.api.ServiceLocator;
 import org.glassfish.hk2.api.ServiceLocatorFactory;
-import org.glassfish.hk2.extras.events.internal.DefaultTopicDistributionService;
 import org.glassfish.hk2.runlevel.RunLevelController;
-import org.glassfish.hk2.utilities.BuilderHelper;
-import org.glassfish.hk2.utilities.ClasspathDescriptorFileFinder;
-import org.glassfish.hk2.utilities.DuplicatePostProcessor;
-import org.glassfish.hk2.utilities.ServiceLocatorUtilities;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
 import java.time.Duration;
 import java.util.List;
 import java.util.Optional;
@@ -49,32 +39,6 @@ class StrideApplicationImpl implements StrideApplication {
         BeanDiscovery beanDiscovery = new BeanDiscovery(this.configuration, serviceLocator);
         this.lifecycle = new Lifecycle(this.configuration, serviceLocator, beanDiscovery);
         //log.debug("Config:\n{}", this.configuration.toPrettyString());
-    }
-
-    @Override
-    public ServiceLocator getServiceLocator() {
-        return serviceLocator;
-    }
-
-    @Override
-    public Optional<ServerConnector> getConnector() {
-        Server server = serviceLocator.getService(Server.class);
-        if (server == null) {
-            int runLevel = ofNullable(serviceLocator.getService(RunLevelController.class))
-                    .map(RunLevelController::getCurrentRunLevel)
-                    .orElse(-1);
-            log.error("Jetty Server is NOT started (run-level: {})", runLevel);
-            return Optional.empty();
-        }
-        for (Connector connector : server.getConnectors()) {
-            // the first connector should be the http connector
-            ServerConnector serverConnector = (ServerConnector) connector;
-            List<String> protocols = serverConnector.getProtocols();
-            if (!protocols.contains("ssl") && (protocols.contains("http/1.1") || protocols.contains("h2c"))) {
-                return Optional.of(serverConnector);
-            }
-        }
-        return Optional.empty();
     }
 
     @Override
@@ -109,80 +73,32 @@ class StrideApplicationImpl implements StrideApplication {
         stop();
     }
 
-    static class BeanDiscovery {
-        private final ApplicationConfiguration configuration;
-        private final ServiceLocator serviceLocator;
+    // ---------------------------------------------------------------------------------------------------------------
 
-        BeanDiscovery(ApplicationConfiguration configuration, ServiceLocator serviceLocator) {
-            this.configuration = configuration;
-            this.serviceLocator = serviceLocator;
-            populateServiceLocator();
-        }
-
-        void populateServiceLocator() throws MultiException {
-            DynamicConfigurationService dcs = serviceLocator.getService(DynamicConfigurationService.class);
-
-            DynamicConfiguration dynamicConfiguration = ServiceLocatorUtilities.createDynamicConfiguration(serviceLocator);
-            dynamicConfiguration.addActiveDescriptor(BuilderHelper.createConstantDescriptor(configuration));
-            dynamicConfiguration.addActiveDescriptor(BuilderHelper.createConstantDescriptor(this));
-            dynamicConfiguration.addActiveDescriptor(DefaultTopicDistributionService.class);
-            dynamicConfiguration.commit();
-
-            Populator populator = dcs.getPopulator();
-
-            try {
-                populator.populate(
-                        new ClasspathDescriptorFileFinder()
-                        , new DuplicatePostProcessor()
-                        , new DefaultConfigurationPostPopulatorProcessor(configuration)
-                );
-            } catch (IOException e) {
-                throw new MultiException(e);
-            }
-        }
+    @Override
+    public ServiceLocator getServiceLocator() {
+        return serviceLocator;
     }
 
-    static class Lifecycle {
-        private final ApplicationConfiguration configuration;
-        private final ServiceLocator serviceLocator;
-        private final BeanDiscovery beanDiscovery;
-
-        Lifecycle(ApplicationConfiguration configuration, ServiceLocator serviceLocator, BeanDiscovery beanDiscovery) {
-            this.configuration = configuration;
-            this.serviceLocator = serviceLocator;
-            this.beanDiscovery = beanDiscovery;
+    @Override
+    public Optional<ServerConnector> getConnector() {
+        Server server = serviceLocator.getService(Server.class);
+        if (server == null) {
+            int runLevel = ofNullable(serviceLocator.getService(RunLevelController.class))
+                    .map(RunLevelController::getCurrentRunLevel)
+                    .orElse(-1);
+            log.error("Jetty Server is NOT started (run-level: {})", runLevel);
+            return Optional.empty();
         }
-
-        void configure() {
-            RunLevelController runLevelController = serviceLocator.getService(RunLevelController.class);
-            runLevelController.setThreadingPolicy(RunLevelController.ThreadingPolicy.valueOf(configuration.asString("hk2.threadpolicy", "FULLY_THREADED")));
-            runLevelController.setMaximumUseableThreads(configuration.asInt("hk2.threadcount", 20));
-        }
-
-        void preStart() {
-            beanDiscovery.populateServiceLocator();
-
-            configure();
-
-            RunLevelController runLevelController = serviceLocator.getService(RunLevelController.class);
-            if (runLevelController.getMaximumUseableThreads() > 3) {
-                runLevelController.proceedTo(3); // start run levels before jetty server start
+        for (Connector connector : server.getConnectors()) {
+            // the first connector should be the http connector
+            ServerConnector serverConnector = (ServerConnector) connector;
+            List<String> protocols = serverConnector.getProtocols();
+            if (!protocols.contains("ssl") && (protocols.contains("http/1.1") || protocols.contains("h2c"))) {
+                return Optional.of(serverConnector);
             }
         }
-
-        void start() {
-            RunLevelController runLevelController = serviceLocator.getService(RunLevelController.class);
-            runLevelController.proceedTo(runLevelController.getMaximumUseableThreads()); // start all run levels
-        }
-
-        void preStop() {
-            RunLevelController runLevelController = serviceLocator.getService(RunLevelController.class);
-            runLevelController.proceedTo(0);
-        }
-
-        void stop() {
-            serviceLocator.shutdown();
-        }
+        return Optional.empty();
     }
 
     // ---------------------------------------------------------------------------------------------------------------
